requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms>=2.3
crispy-bootstrap5==2024.10
python-dotenv==1.0.1
django-cleanup==8.1.0
django-widget-tweaks==1.5.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg[binary]==3.1.18
django-filter==24.1
